Move ingestion helpers to common
