+++ src/ingest/snodas_download.py
@@ -4,6 +4,7 @@ import requests
 import tarfile
 import gzip
 import logging
+from ingest.common import setup_logging, format_date, build_output_dir
 from datetime import datetime
 from pathlib import Path
 
@@ -16,49 +17,24 @@ TARGET_DATE = datetime.strptime(args.date, "%Y-%m-%d") if args.date else datetim
 BASE_DIR = Path("data/snodas")
 BASE_URL = "https://noaadata.apps.nsidc.org/NOAA/G02158/unmasked"
 
-def setup_logging(name: str, date: str):
-    logs_dir = Path("logs")
-    logs_dir.mkdir(exist_ok=True)
-
-    log_path = f"logs/{name}_{date}.log"
-
-    logging.basicConfig(
-        level=logging.INFO,
-        format="%(asctime)s %(levelname)s %(message)s",
-        handlers=[
-            logging.FileHandler(log_path),
-            logging.StreamHandler() # also logs to console
-        ]
-    )
-
 setup_logging("snodas", args.date)
 
 # ========== Utility Functions ==========
-
-def format_date(date: datetime):
-    return date.strftime("%Y"), date.strftime("%m_%b"), date.strftime("%Y%m%d"), date.strftime("%d")
 
 def build_snodas_url(date: datetime):
     """
     Builds the SNODAS tar file URL and expected filename for the given date.
     """
-    year, month, file_date, _ = format_date(date)
+    year, _, _, month_str, file_date = format_date(date)
     filename = f"SNODAS_unmasked_{file_date}.tar"
-    url = f"{BASE_URL}/{year}/{month}/{filename}"
+    url = f"{BASE_URL}/{year}/{month_str}/{filename}"
     return url, filename
-
-def build_output_dir(date: datetime) -> Path:
-    """
-    Builds the local directory path based on date.
-    """
-    year, month, _, day = format_date(date)
-    return BASE_DIR / year / month / day
 
 # ========== Main Download Logic ==========
 
 def download_snodas_file(date: datetime):
     url, filename = build_snodas_url(date)
-    output_dir = build_output_dir(date)
+    output_dir = build_output_dir(date, BASE_DIR)
     local_tar_path = output_dir / filename
 
     logging.info("Starting SNODAS ingest for %s", date.strftime('%Y-%m-%d'))
